somme mod

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: elel-yak <elel-yak@student.1337.ma>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/08 15:13:07 by elel-yak          #+#    #+#             */
-/*   Updated: 2022/10/08 22:38:03 by elel-yak         ###   ########.fr       */
+/*   Updated: 2022/10/22 14:21:00 by elel-yak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
